requirements.txt
asgiref==3.4.1
cffi==1.14.6
cryptography==3.4.7
Django==3.2.5
django-cors-headers==3.7.0
django-rest-framework==0.1.0
django-rest-knox==4.1.0
django-rest-passwordreset==1.2.0
djangorestframework==3.12.4
pycparser==2.20
pytz==2021.1
sqlparse==0.4.1
